refactor: 분리된 객체를 사용하도록 수정
