+++ src/main/java/step3/domain/CarRacingScoreboard.java
@@ -8,15 +8,15 @@ import java.util.stream.Collectors;
 
 public class CarRacingScoreboard {
 
-    private List<List<Integer>> scoreboard = new ArrayList<>();
+    private static List<List<Integer>> scoreboard = new ArrayList<>();
 
-    public void recordScoreboard(List<Car>cars) {
+    public static void recordScoreboard(List<Car>cars) {
         scoreboard.add(cars.stream()
                 .map(Car::getPosition)
                 .collect(Collectors.toList()));
     }
 
-    public List<List<Integer>> getScoreboard() {
+    public static List<List<Integer>> getScoreboard() {
         return scoreboard;
     }
